+++ CHANGES.rst
@@ -7,6 +7,7 @@ master (unreleased)
 - Fix `FieldTracker.has_changed()` and `FieldTracker.previous()` to return
   correct responses for deferred fields.
 - Support `reversed` for all kinds of `Choices` objects, fixes GH-309
+- Fix Model instance non picklable GH-330
 
 3.1.2 (2018.05.09)
 ------------------